backend/requirements-test.txt
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-cov==4.1.0
requests-mock==1.11.0
factory-boy==3.3.0
faker==20.1.0